app/requirements.txt
fastapi
uvicorn
openai
azure-identity
azure-search-documents>=11.4.0b7
python-dotenv
httpx
pydantic 